+++ src/main/java/com/ururulab/ururu/groupBuy/dto/request/GroupBuyOptionRequest.java
@@ -1,4 +1,4 @@
-package com.ururulab.ururu.groupBuy.controller.dto.request;
+package com.ururulab.ururu.groupBuy.dto.request;
 
 import com.ururulab.ururu.groupBuy.domain.entity.GroupBuy;
 import com.ururulab.ururu.groupBuy.domain.entity.GroupBuyOption;
@@ -6,7 +6,7 @@ import com.ururulab.ururu.product.domain.entity.ProductOption;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
-import static com.ururulab.ururu.groupBuy.controller.dto.validation.GroupBuyValidationMessages.*;
+import static com.ururulab.ururu.groupBuy.dto.validation.GroupBuyValidationMessages.*;
 
 public record GroupBuyOptionRequest(
         Long id, // Update시에만 사용 (Create시에는 null)
